+++ packages/codegloss/src/index.ts
@@ -2,6 +2,7 @@ import { defineCodeGloss } from './core/code-gloss.element';
 
 export { CodeGlossElement, defineCodeGloss } from './core/code-gloss.element';
 export { setDefaultHighlighter } from './core/default-highlighter.helpers';
+export { initCodegloss } from './core/init-codegloss.helpers';
 
 export type {
 	Annotation,